refactor(ingest): extract HTTP fetch dispatch to wiki/fetch.py
